+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/action/SubtaskTemplateImportAction.java
@@ -16,13 +16,13 @@ import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskTemplate;
 import de.codescape.jira.plugins.multiplesubtasks.model.ShowSubtaskTemplate;
 import de.codescape.jira.plugins.multiplesubtasks.model.Subtask;
 import de.codescape.jira.plugins.multiplesubtasks.service.SubtaskTemplateService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 
+import javax.inject.Inject;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -56,7 +56,7 @@ public class SubtaskTemplateImportAction extends JiraWebActionSupport {
     private final StringWriter output = new StringWriter();
     private boolean quickSubtasksCanImport;
 
-    @Autowired
+    @Inject
     public SubtaskTemplateImportAction(@ComponentImport UserSearchService userSearchService,
                                        @ComponentImport PluginSettingsFactory pluginSettingsFactory,
                                        @ComponentImport ProjectManager projectManager,