+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/rest/entities/CreatedSubtaskEntity.java
@@ -1,9 +1,9 @@
 package de.codescape.jira.plugins.multiplesubtasks.rest.entities;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import de.codescape.jira.plugins.multiplesubtasks.model.CreatedSubtask;
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-import org.codehaus.jackson.annotate.JsonProperty;
-import org.codehaus.jackson.annotate.JsonPropertyOrder;
 
 import java.util.List;