+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/service/SubtaskTemplateService.java
@@ -9,9 +9,9 @@ import de.codescape.jira.plugins.multiplesubtasks.model.SubtaskTemplateType;
 import de.codescape.jira.plugins.multiplesubtasks.model.TemplateSortOrder;
 import net.java.ao.DBParam;
 import net.java.ao.Query;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import javax.inject.Inject;
 import java.util.Arrays;
 import java.util.List;
 
@@ -25,7 +25,7 @@ public class SubtaskTemplateService {
     private final ActiveObjects activeObjects;
     private final MultipleSubtasksConfigurationService multipleSubtasksConfigurationService;
 
-    @Autowired
+    @Inject
     public SubtaskTemplateService(@ComponentImport ActiveObjects activeObjects,
                                   MultipleSubtasksConfigurationService multipleSubtasksConfigurationService) {
         this.activeObjects = activeObjects;