+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/upgrade/Upgrade01MaximumTemplatesSettings.java
@@ -3,10 +3,10 @@ package de.codescape.jira.plugins.multiplesubtasks.upgrade;
 import com.atlassian.plugin.spring.scanner.annotation.export.ExportAsService;
 import com.atlassian.sal.api.upgrade.PluginUpgradeTask;
 import de.codescape.jira.plugins.multiplesubtasks.service.MultipleSubtasksConfigurationService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nonnull;
+import javax.inject.Inject;
 
 /**
  * Initialize configuration with default settings for maximum templates per user and per project.
@@ -19,7 +19,7 @@ public class Upgrade01MaximumTemplatesSettings extends AbstractUpgradeTask {
 
     private final MultipleSubtasksConfigurationService multipleSubtasksConfigurationService;
 
-    @Autowired
+    @Inject
     public Upgrade01MaximumTemplatesSettings(MultipleSubtasksConfigurationService multipleSubtasksConfigurationService) {
         this.multipleSubtasksConfigurationService = multipleSubtasksConfigurationService;
     }